+++ deps.ts
@@ -5,10 +5,8 @@ export { default as fast_json_stable_stringify } from "https://esm.sh/fast-json-
 export * as rfc4648 from "https://esm.sh/rfc4648@1.5.0";
 export * as sha256_uint8array from "https://esm.sh/sha256-uint8array@0.10.3";
 export { Superbus } from "./src/superbus/superbus.ts";
-export { Simplebus } from "./src/superbus/simplebus.ts";
 export { SuperbusMap } from "./src/superbus_map/superbus_map.ts";
 export * as ed from "https://raw.githubusercontent.com/sgwilym/noble-ed25519/7af9329476ff2f2a0e524a9f78e36d09704efc63/mod.ts";
-export { Lock } from "https://cdn.skypack.dev/concurrency-friends@5.2.0?dts";
 export {
   type IConnection,
   type ITransport,